+++ src/Avalonia.Animation/KeySpline.cs
@@ -340,20 +340,4 @@ namespace Avalonia.Animation
             }
         }
     }
-
-    /// <summary>
-    /// Converts string values to <see cref="KeySpline"/> values
-    /// </summary>
-    public class KeySplineTypeConverter : TypeConverter
-    {
-        public override bool CanConvertFrom(ITypeDescriptorContext context, Type sourceType)
-        {
-            return sourceType == typeof(string);
-        }
-
-        public override object ConvertFrom(ITypeDescriptorContext context, CultureInfo culture, object value)
-        {
-            return KeySpline.Parse((string)value, culture);
-        }
-    }
 }
